requirements/base.txt


# Bleeding edge Django
django==1.10.2

# Configuration
django-environ==0.4.0
whitenoise==3.2.2


# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0

# Models
django-model-utils==2.6

# Images
Pillow==3.4.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.27.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.7

# Redis support
django-redis==4.5.0
redis>=2.10.5

# Django REST Framework
djangorestframework==3.4.7

django-rest-auth==0.8.1
